# gen/openapiv2/gortexa.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ai/v1/annotations.proto",
    "description": "Package ai.v1 defines annotations that expose gRPC methods and request\nfields to AI/LLM agents. They are opt-in: a method is only surfaced as a\ntool when ai_tool.expose is true.",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "ResourceService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/resources": {
      "get": {
        "summary": "ListResources lists resources, optionally filtered by owner. Read-only tool.",
        "operationId": "ResourceService_ListResources",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListResourcesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googleRpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pageToken",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "owner",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "ResourceService"
        ]
      },
      "post": {
        "summary": "CreateResource creates a resource. Mutating tool.",
        "operationId": "ResourceService_CreateResource",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1Resource"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googleRpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "resource",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1Resource"
            }
          }
        ],
        "tags": [
          "ResourceService"
        ]
      }
    },
    "/v1/resources/{id}": {
      "get": {
        "summary": "GetResource fetches a single resource by id. Read-only tool.",
        "operationId": "ResourceService_GetResource",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1Resource"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googleRpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "ResourceService"
        ]
      },
      "delete": {
        "summary": "DeleteResource permanently deletes a resource. Destructive tool.",
        "operationId": "ResourceService_DeleteResource",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googleRpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "ResourceService"
        ]
      },
      "patch": {
        "summary": "UpdateResource partially updates a resource (PATCH): fields present in the\nrequest replace the stored values; omitted fields are left untouched.",
        "operationId": "ResourceService_UpdateResource",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1Resource"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googleRpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ResourceServiceUpdateResourceBody"
            }
          }
        ],
        "tags": [
          "ResourceService"
        ]
      }
    }
  },
  "definitions": {
    "ResourceServiceUpdateResourceBody": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "owner": {
          "type": "string"
        },
        "status": {
          "$ref": "#/definitions/resourceV1Status"
        }
      },
      "description": "UpdateResourceRequest is a PATCH body: id selects the row, every mutable field\nis `optional` so its validation rule fires only when the caller sets it. A\npartial update that omits a field passes validation and leaves that field\nuntouched — reusing Resource here would force every field's create-time rule\n(e.g. name min_len) onto every partial update."
    },
    "googleRpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "resourceV1Status": {
      "type": "string",
      "enum": [
        "STATUS_UNSPECIFIED",
        "STATUS_ACTIVE",
        "STATUS_ARCHIVED"
      ],
      "default": "STATUS_UNSPECIFIED",
      "description": "Status is the lifecycle state of a Resource. Enums downgrade to string in\nevery AI-provider schema."
    },
    "v1ListResourcesResponse": {
      "type": "object",
      "properties": {
        "resources": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Resource"
          }
        },
        "nextPageToken": {
          "type": "string"
        }
      }
    },
    "v1Resource": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "owner": {
          "type": "string"
        },
        "status": {
          "$ref": "#/definitions/resourceV1Status"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "description": "Resource is the entity managed by ResourceService."
    }
  }
}